% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp_typ.R
\name{build_latent_space}
\alias{build_latent_space}
\title{Build Latent Response Types Space}
\usage{
build_latent_space(k, l = NULL, assumptions = NULL)
}
\description{
Build Latent Response Types Space
}
\keyword{internal}
