% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_prep_probs.R
\name{prep_empirical_probs}
\alias{prep_empirical_probs}
\title{Prepare Empirical Probabilities}
\usage{
prep_empirical_probs(
  data,
  outcome,
  treatment,
  group,
  k = NULL,
  m = NULL,
  l = NULL
)
}
\description{
Prepare Empirical Probabilities
}
