% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_ACE.R
\name{solve_ace_bound}
\alias{solve_ace_bound}
\title{Solve a Single ACE Bound via Column Generation}
\usage{
solve_ace_bound(
  target_a,
  target_b,
  is_upper_bound,
  slack_val,
  tolerance,
  max_iter,
  types_Y,
  prep_data,
  verbose
)
}
\description{
Solve a Single ACE Bound via Column Generation
}
\keyword{internal}
